create AmpereMultipleMatmulScheduler
